Set up boilerplates, set up functions, set up IO
Write introduction of the project on README, as well as set expectations
for the project's completion.

Initialize boilerplate HTML and JavaScript files.

Initialize computerPlay() function to determine the computer's choice
using a random number generator with the floor() and random() functions.
That integer will be in the range of 0-2 inclusive and it will then be
run through conditionals to determine which choice that integer
represents.

Initialize playerPlay() function to parse player choice using a prompt()
function and convert it into a consistent uppercase string using the
toUpperCase() method.

Initialize playRound() function to compare playerSelection and
computerSelection using a chain of conditionals that alert to the
result of the round, then return that result for future use.

Initialize game() method to keep track of the player and computer's
scores, as well as loop through the playRound() function 5 times, add
to their respective score when applicable. Utilize conditionals in
game() function to announce the overall winner of the 5 rounds.